Add enqueue. Rework DB to user_id to BigInt

+++ project_root/model/queue_entry.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, UniqueConstraint
+from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, UniqueConstraint, BigInteger
 from sqlalchemy.orm import declarative_base, relationship
 
 from project_root.model import Base
@@ -9,7 +9,7 @@ class QueueEntry(Base):
 
     id = Column(Integer, primary_key=True)
     queue_id = Column(Integer, ForeignKey('queues.id'), nullable=False)
-    user_id = Column(Integer, ForeignKey('users.id'), nullable=False)
+    user_id = Column(BigInteger, ForeignKey('users.id'), nullable=False)
     place = Column(Integer, nullable=False)
     timestamp = Column(DateTime, nullable=False)
